Issue 27: Moved "common" Jackson code into a separate module
