+++ common/cxf/BUCK
@@ -5,7 +5,7 @@ java_library(
         "//lib/spring:spring-context",
     ],
     visibility = [
-        "//core/jackson:queue-triage-core-jackson",
+        "//common/jackson:queue-triage-common-jackson",
         "//core/server:queue-triage-core-server",
     ]
 )
@@ -21,9 +21,9 @@ java_library(
     deps = [
         ":queue-triage-cxf-client",
         "//lib/apache-cxf:cxf-rt-transports-http-jetty",
+        "//lib/javax:javax-servlet-api",
         "//lib/spring:spring-boot",
         "//lib/spring:spring-context",
-        "//lib/javax:javax-servlet-api",
     ],
     visibility = [
         "//core/component-test:test",
