+++ common/jackson/src/main/java/uk/gov/dwp/queue/triage/jackson/ISO8601DateFormatWithMilliSeconds.java
@@ -1,4 +1,4 @@
-package uk.gov.dwp.queue.triage.core.jackson;
+package uk.gov.dwp.queue.triage.jackson;
 
 import com.fasterxml.jackson.databind.util.ISO8601DateFormat;
 import com.fasterxml.jackson.databind.util.ISO8601Utils;
