+++ common/jackson/src/main/java/uk/gov/dwp/queue/triage/jackson/configuration/JacksonConfiguration.java
@@ -1,4 +1,4 @@
-package uk.gov.dwp.queue.triage.core.jackson.configuration;
+package uk.gov.dwp.queue.triage.jackson.configuration;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -10,10 +10,10 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import uk.gov.dwp.migration.mongo.demo.cxf.client.CxfConfiguration;
-import uk.gov.dwp.queue.triage.core.jackson.ISO8601DateFormatWithMilliSeconds;
-import uk.gov.dwp.queue.triage.core.jackson.IdDeserializer;
-import uk.gov.dwp.queue.triage.core.jackson.IdSerializer;
 import uk.gov.dwp.queue.triage.id.Id;
+import uk.gov.dwp.queue.triage.jackson.ISO8601DateFormatWithMilliSeconds;
+import uk.gov.dwp.queue.triage.jackson.IdDeserializer;
+import uk.gov.dwp.queue.triage.jackson.IdSerializer;
 
 @Configuration
 @Import({
