+++ core/component-test/BUCK
@@ -7,13 +7,13 @@ java_library(
     resources_root = "src/main/resources",
     deps = [
         '//common/id:queue-triage-common-id',
+        '//common/jackson:queue-triage-common-jackson',
         '//common/jgiven:jgiven',
         '//core/activemq:queue-triage-core-activemq',
         '//core/client:queue-triage-core-client',
         '//core/client-test-support:queue-triage-core-client-test-support',
         '//core/dao-mongo-test-support:queue-triage-core-dao-mongo-test-support',
         '//core/domain:queue-triage-core-domain',
-        '//core/jackson:queue-triage-core-jackson',
         '//core/server:queue-triage-core-server',
         '//lib/activemq:activemq-client',
         '//lib/apache-cxf:cxf-core',
