+++ core/dao-mongo/BUCK
@@ -3,9 +3,9 @@ java_library(
     srcs = glob(["src/main/java/**/*.java"]),
     deps = [
         "//common/id:queue-triage-common-id",
+        '//common/jackson:queue-triage-common-jackson-spring',
         '//core/dao:queue-triage-core-dao',
         '//core/domain:queue-triage-core-domain',
-        '//core/jackson:queue-triage-core-jackson',
         '//lib/jackson:jackson-databind',
         '//lib/mongo:mongodb-driver',
         '//lib/spring:spring-boot',
