--- core/jackson/BUCK
@@ -1,15 +1,0 @@
-java_library(
-    name = "queue-triage-core-jackson",
-    srcs = glob(["src/main/java/**/*.java"]),
-    deps =[
-        "//common/cxf:queue-triage-cxf-client",
-        "//common/id:queue-triage-common-id",
-        "//lib/javax:javax-ws-rs-api",
-        '//lib/jackson:jackson-datatype-jdk8',
-        '//lib/jackson:jackson-jaxrs-json-provider',
-        '//lib/spring:spring-context',
-    ],
-    visibility = [
-        "PUBLIC"
-    ]
-)
