+++ core/server/BUCK
@@ -15,12 +15,12 @@ java_library(
         '//common/cxf:queue-triage-cxf-client',
         '//common/health:queue-triage-health',
         '//common/id:queue-triage-common-id',
+        '//common/jackson:queue-triage-common-jackson-spring',
         '//core/activemq:queue-triage-core-activemq',
         '//core/client:queue-triage-core-client',
         '//core/dao:queue-triage-core-dao',
         '//core/dao-mongo:queue-triage-core-dao-mongo',
         '//core/domain:queue-triage-core-domain',
-        '//core/jackson:queue-triage-core-jackson',
         '//core/jms:queue-triage-core-jms',
         '//core/search:queue-triage-core-search',
         '//core/search:queue-triage-core-search-mongo',
